% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_current_quarter}
\alias{get_current_quarter}
\title{The Current Quarter}
\usage{
get_current_quarter()
}
\value{
a numeric value containing the current quarter.
}
\description{
Returns the current quarter.
}
\author{
Sami Kallinen
}

