% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_current_week}
\alias{get_current_week}
\title{The Current Week number}
\usage{
get_current_week()
}
\value{
a number with the current week
}
\description{
The current week number
}
\author{
Sami Kallinen
}

