% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_current_year}
\alias{get_current_year}
\title{The Current Year}
\usage{
get_current_year()
}
\value{
a numeric value containing the current year
}
\description{
Returns the current year.
}
\author{
Sami Kallinen
}

