% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_dummy_weeks}
\alias{get_dummy_weeks}
\title{Dummy Week Data Frame}
\usage{
get_dummy_weeks()
}
\value{
a data.frame
}
\description{
A dummy data.frame with week numbers, month and year for the current year.
}
\author{
Sami Kallinen
}

