% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_last_quarter_week}
\alias{get_last_quarter_week}
\title{The last week number of current quarter}
\usage{
get_last_quarter_week()
}
\value{
a number
}
\description{
The last week number of current quarter
}
\author{
Sami Kallinen
}

