% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_period_start_monday}
\alias{get_period_start_monday}
\title{Date of the Monday when This Quarter Started}
\usage{
get_period_start_monday()
}
\value{
a POSIXct date
}
\description{
Date of the starting monday of the week when this quarter started.
}
\author{
Sami Kallinen
}

