% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_quarter_dates}
\alias{get_quarter_dates}
\title{The Quarter Start and End dates}
\usage{
get_quarter_dates()
}
\value{
a data.frame with start dates and end dates for all
}
\description{
Data frame with quarter dates.
}
\author{
Sami Kallinen
}

