% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_quarter_duration_weeks}
\alias{get_quarter_duration_weeks}
\title{Duration of Current Quarter}
\usage{
get_quarter_duration_weeks()
}
\value{
a number
}
\description{
The duration of current quarter in weeks
}
\author{
Sami Kallinen
}

