% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_quarter_end}
\alias{get_quarter_end}
\title{Last Day of Current Quarter}
\usage{
get_quarter_end()
}
\value{
the last day of the quarter in POSIXct
}
\description{
The date when current quarter ends
}
\author{
Sami Kallinen
}

