% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_quarter_start}
\alias{get_quarter_start}
\title{First Day of Current Quarter}
\usage{
get_quarter_start()
}
\value{
the first day of the quarter in POSIXct
}
\description{
The date when current quarter started
}
\author{
Sami Kallinen
}

