% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_time_seq}
\alias{get_time_seq}
\title{Sequence the Dates in Current Quarter}
\usage{
get_time_seq()
}
\value{
a sequence of POSIXct dates
}
\description{
Produces the sequence of dates in the current quarter
}
\author{
Sami Kallinen
}

