% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_time_zone}
\alias{get_time_zone}
\title{Stores The Default Time Zone}
\usage{
get_time_zone()
}
\value{
a string containing "EET" eastern european time code.
}
\description{
Rerturns the default time zone.
}
\author{
Sami Kallinen
}

