% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{get_today_string}
\alias{get_today_string}
\title{Today as a String}
\usage{
get_today_string()
}
\value{
a string
}
\description{
Returns today's date as a string
}
\author{
Sami Kallinen
}

