% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{\%+\%}
\alias{\%+\%}
\title{\%+\% Operator Infix Function For String Concatenation}
\usage{
x \%+\% y
}
\arguments{
\item{x}{strings}

\item{y}{strings}
}
\value{
a concatenated string
}
\description{
Add string operation
}
\author{
Sami Kallinen
}

