% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{seq_date}
\alias{seq_date}
\title{Sequence Dates}
\usage{
seq_date(start, end)
}
\arguments{
\item{start}{POSIXct date}

\item{end}{POSIXct date}
}
\value{
a sequence of POSIXct dates
}
\description{
Sequence Dates
}
\author{
Sami Kallinen
}

