% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date-time-helpers.R
\name{trunc_date}
\alias{trunc_date}
\title{Trunctates a POSIXct date}
\usage{
trunc_date(date, units)
}
\arguments{
\item{date}{a POSICct date object}

\item{units}{one of the following units c\("secs", "mins", "hours", "days"\). Can be abbreviated.}
}
\value{
a POSIXct object
}
\description{
Does the same as truc but for a POSIXct date.
}
\author{
Sami Kallinen
}

